Move `AlgorithmIdentifier` type into `alg_id`
